add ILogger interface
